src/chatbot-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pymavlink==2.4.39
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.4
google-generativeai==0.5.2
structlog==24.1.0
python-multipart==0.0.9
duckdb==0.10.2
pandas==2.2.1